Move Refactoring class to refactoring subpackage.
