+++ src/main/java/org/ml_methods_group/algorithm/AKMeans.java
@@ -19,6 +19,7 @@ package org.ml_methods_group.algorithm;
 import org.apache.log4j.Logger;
 import org.ml_methods_group.algorithm.entity.Entity;
 import org.ml_methods_group.algorithm.entity.EntitySearchResult;
+import org.ml_methods_group.algorithm.refactoring.Refactoring;
 import org.ml_methods_group.config.Logging;
 import org.ml_methods_group.utils.AlgorithmsUtil;
