+++ src/main/java/org/ml_methods_group/refactoring/RefactorIntentionAction.java
@@ -26,7 +26,7 @@ import com.intellij.psi.PsiFile;
 import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
-import org.ml_methods_group.algorithm.Refactoring;
+import org.ml_methods_group.algorithm.refactoring.Refactoring;
 import org.ml_methods_group.utils.RefactoringUtil;
 
 import java.util.Collections;
